day-11/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
pydantic>=2.0
python-dotenv>=1.0.0
PyYAML>=6.0
faiss-cpu>=1.8.0
